# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seat Reservation API",
    "description": "\nSells assigned seats for a show and guarantees each seat is sold exactly once, even when\nthousands of buyers want the same seat in the same second.\n\n**Try it from this page**\n\n1. `POST /auth/token` with `{\"user_id\": \"alice\"}`, then copy `access_token`.\n2. Click **Authorize**: paste the token under **BearerAuth**, and the admin key under\n   **AdminKey** (only needed to create shows and run the audit).\n3. `POST /shows` to create a show, then `POST /shows/{show_id}/reserve`.\n4. `GET /shows/{show_id}` shows every seat: `available + held + confirmed == total` always holds.\n\n**Rules worth knowing**\n\n- Identity comes only from the token; a `user_id` in a request body is ignored.\n- Money is integer paise; a float or string price is rejected.\n- A multi-seat request is all-or-nothing.\n- The same idempotency key returns the same reservation (`200` replay); the same key with\n  other seats is `409`.\n- Every error has the same shape, `{\"error\": {\"code\", \"message\", ...}, \"request_id\"}`.\n  Declines are `409`, never `5xx`. If the database is unreachable: `503` with `Retry-After`.\n\nSource, write-up and the one-command burst:\n[github.com/jaymovaliya26/seat-reservation](https://github.com/jaymovaliya26/seat-reservation)\n",
    "version": "1.1.0"
  },
  "paths": {
...
    "/shows/{show_id}/reserve": {
      "post": {
        "tags": [
          "reservations"
        ],
        "summary": "Reserve seats",
        "operationId": "reserve_shows__show_id__reserve_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Show Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[!-~]{1,128}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Makes retries safe. May instead go in the body as idempotency_key.",
              "title": "Idempotency-Key"
            },
            "description": "Makes retries safe. May instead go in the body as idempotency_key."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            }
          },
          "200": {
            "description": "Replay of an earlier request with the same idempotency key (header Idempotent-Replayed: true). Nothing new is booked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token (or missing admin key)",
            "content": {
              "application/json": {
                "examples": {
                  "missing token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Missing bearer token"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "expired token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Token is invalid or expired"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found (including another user's reservation)",
            "content": {
              "application/json": {
                "examples": {
                  "not found": {
                    "value": {
                      "error": {
                        "code": "not_found",
                        "message": "Show not found"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "A clean decline: the request was valid, but the current state says no",
            "content": {
              "application/json": {
                "examples": {
                  "seat_taken": {
                    "value": {
                      "error": {
                        "code": "seat_taken",
                        "message": "Seats already taken: A12",
                        "seats": [
                          "A12"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "per_user_limit": {
                    "value": {
                      "error": {
                        "code": "per_user_limit",
                        "message": "At most 4 seats per user for this show",
                        "limit": 4,
                        "held": 3,
                        "requested": 2
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "idempotency_key_reused": {
                    "value": {
                      "error": {
                        "code": "idempotency_key_reused",
                        "message": "This idempotency key was already used for a different request",
                        "reservation_id": "0b6f2f6e-6a1e-4b8e-9d3a-2c7f1e5a9b10"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "already_cancelled": {
                    "value": {
                      "error": {
                        "code": "already_cancelled",
                        "message": "Reservation is already cancelled"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The request is malformed",
            "content": {
              "application/json": {
                "examples": {
                  "invalid body": {
                    "value": {
                      "error": {
                        "code": "invalid_request",
                        "message": "The request is not valid",
                        "problems": [
                          {
                            "field": "body.price_paise",
                            "problem": "Input should be a valid integer"
                          }
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "unknown seats": {
                    "value": {
                      "error": {
                        "code": "unknown_seats",
                        "message": "Seats do not exist in this show: Z9",
                        "seats": [
                          "Z9"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable or overloaded; retry with the same key after `Retry-After` seconds",
            "content": {
              "application/json": {
                "examples": {
                  "unavailable": {
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "message": "The database is unreachable; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "busy": {
                    "value": {
                      "error": {
                        "code": "database_busy",
                        "message": "The database is overloaded; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/reservations/{reservation_id}": {
      "get": {
        "tags": [
          "reservations"
        ],
        "summary": "View your reservation",
        "description": "Your own reservation. Someone else's is a 404, so its existence is never revealed.",
        "operationId": "get_reservation_reservations__reservation_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token (or missing admin key)",
            "content": {
              "application/json": {
                "examples": {
                  "missing token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Missing bearer token"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "expired token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Token is invalid or expired"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found (including another user's reservation)",
            "content": {
              "application/json": {
                "examples": {
                  "not found": {
                    "value": {
                      "error": {
                        "code": "not_found",
                        "message": "Show not found"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The request is malformed",
            "content": {
              "application/json": {
                "examples": {
                  "invalid body": {
                    "value": {
                      "error": {
                        "code": "invalid_request",
                        "message": "The request is not valid",
                        "problems": [
                          {
                            "field": "body.price_paise",
                            "problem": "Input should be a valid integer"
                          }
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "unknown seats": {
                    "value": {
                      "error": {
                        "code": "unknown_seats",
                        "message": "Seats do not exist in this show: Z9",
                        "seats": [
                          "Z9"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable or overloaded; retry with the same key after `Retry-After` seconds",
            "content": {
              "application/json": {
                "examples": {
                  "unavailable": {
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "message": "The database is unreachable; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "busy": {
                    "value": {
                      "error": {
                        "code": "database_busy",
                        "message": "The database is overloaded; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/reservations/{reservation_id}/cancel": {
      "post": {
        "tags": [
          "reservations"
        ],
        "summary": "Cancel your reservation",
        "description": "Cancel your own reservation. Its seats become available to anyone at once.",
        "operationId": "cancel_reservations__reservation_id__cancel_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token (or missing admin key)",
            "content": {
              "application/json": {
                "examples": {
                  "missing token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Missing bearer token"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "expired token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Token is invalid or expired"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found (including another user's reservation)",
            "content": {
              "application/json": {
                "examples": {
                  "not found": {
                    "value": {
                      "error": {
                        "code": "not_found",
                        "message": "Show not found"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "A clean decline: the request was valid, but the current state says no",
            "content": {
              "application/json": {
                "examples": {
                  "seat_taken": {
                    "value": {
                      "error": {
                        "code": "seat_taken",
                        "message": "Seats already taken: A12",
                        "seats": [
                          "A12"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "per_user_limit": {
                    "value": {
                      "error": {
                        "code": "per_user_limit",
                        "message": "At most 4 seats per user for this show",
                        "limit": 4,
                        "held": 3,
                        "requested": 2
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "idempotency_key_reused": {
                    "value": {
                      "error": {
                        "code": "idempotency_key_reused",
                        "message": "This idempotency key was already used for a different request",
                        "reservation_id": "0b6f2f6e-6a1e-4b8e-9d3a-2c7f1e5a9b10"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "already_cancelled": {
                    "value": {
                      "error": {
                        "code": "already_cancelled",
                        "message": "Reservation is already cancelled"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The request is malformed",
            "content": {
              "application/json": {
                "examples": {
                  "invalid body": {
                    "value": {
                      "error": {
                        "code": "invalid_request",
                        "message": "The request is not valid",
                        "problems": [
                          {
                            "field": "body.price_paise",
                            "problem": "Input should be a valid integer"
                          }
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "unknown seats": {
                    "value": {
                      "error": {
                        "code": "unknown_seats",
                        "message": "Seats do not exist in this show: Z9",
                        "seats": [
                          "Z9"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable or overloaded; retry with the same key after `Retry-After` seconds",
            "content": {
              "application/json": {
                "examples": {
                  "unavailable": {
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "message": "The database is unreachable; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "busy": {
                    "value": {
                      "error": {
                        "code": "database_busy",
                        "message": "The database is overloaded; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
...
      "ReservationOut": {
        "properties": {
          "reservation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reservation Id"
          },
          "show_id": {
            "type": "string",
            "format": "uuid",
            "title": "Show Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "seats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seats"
          },
          "amount_paise": {
            "type": "integer",
            "title": "Amount Paise"
          },
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "reservation_id",
          "show_id",
          "user_id",
          "seats",
          "amount_paise",
          "status",
          "created_at"
        ],
        "title": "ReservationOut"
      },
      "ReserveRequest": {
        "properties": {
          "seats": {
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,15}$"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Seats"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[!-~]{1,128}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "seats"
        ],
        "title": "ReserveRequest",
        "examples": [
          {
            "idempotency_key": "order-7f3a",
            "seats": [
              "A12"
            ]
          }
        ]
      },
...
    },
    "securitySchemes": {
      "AdminKey": {
        "type": "apiKey",
        "description": "Admin key, needed only to create shows and run the reconcile audit.",
        "in": "header",
        "name": "X-Admin-Key"
      },
      "BearerAuth": {
        "type": "http",
        "description": "A token from POST /auth/token. Paste only the token; Swagger adds 'Bearer '.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}